src/BCR.BrailleTranscriptor/requirements.txt
pip==22.2.1
werkzeug==2.1.2
Flask==2.1.2
flask-restx==0.5.1
flask_cors==3.0.10
opencensus-ext-azure==1.1.7